+++ minefield/ECommerce/Pricing/NormalPricingStrategy.cs
@@ -6,7 +6,7 @@ namespace minefield.ECommerce.Pricing
 {
 	public class NormalPricingStrategy : IPricingStrategy
 	{
-		public int GetPrice(IList<Product> products)
+		public decimal GetPrice(IList<Product> products)
 		{
 			return products.Select(p => p.Price)
 				           .Sum();
